fix : 피드백 반영
service 계층에 createFixedVoucher, createPercentVoucher를 createVoucher 메서드로 합치기
Voucher 도메인에 Serializable을 선언하는게 아니라 직렬화용 DTO 객체 따로 만들어 처리하기
Test에서 BeforeEach, AfterEach 둘 중 하나만 사용하기

+++ src/main/java/org/prgrms/springbootbasic/entity/FixedAmountVoucher.java
@@ -4,14 +4,14 @@ import java.util.UUID;
 
 public class FixedAmountVoucher extends Voucher {
 
-    private final long amount;
+    private final int amount;
 
-    public FixedAmountVoucher(UUID voucherId, long amount) {
+    public FixedAmountVoucher(UUID voucherId, int amount) {
         super(voucherId);
         this.amount = amount;
     }
 
-    public long getAmount() {
+    public int getAmount() {
         return amount;
     }
 }
